l-tester-master/requirements.txt
aerich==0.9.2
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiomysql==0.3.2
aiosignal==1.4.0
aiosqlite==0.22.1
airtest==1.4.3
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
Appium-Python-Client==5.2.4
APScheduler==3.11.2
async-generator==1.10
asyncclick==8.3.0.7
attrs==25.4.0
autowing==0.6.1
cached-property==1.5.2
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
colored==2.3.1
comtypes==1.4.15
construct==2.10.70
cryptography==46.0.3
dashscope==1.25.8
decorator==5.2.1
Deprecated==1.2.18
deprecation==2.1.0
dictdiffer==0.9.0
distro==1.9.0
facebook-wda==1.5.0
fastapi==0.128.0
fastapi-pagination==0.15.6
ffmpeg-python==0.2.0
filelock==3.20.3
frozenlist==1.8.0
future==1.0.0
greenlet==3.3.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
importlib_metadata==8.7.1
iniconfig==2.3.0
iso8601==2.1.0
Jinja2==3.1.6
jiter==0.12.0
jsonpath==0.82.2
loguru==0.7.3
logzero==1.7.0
MarkupSafe==3.0.3
mss==6.1.0
multidict==6.7.0
mypy_extensions==1.1.0
numpy==1.26.4
openai==1.109.1
opencv-contrib-python==4.6.0.66
outcome==1.3.0.post0
packaging==25.0
pillow==12.1.0
playwright==1.57.0
pluggy==1.6.0
propcache==0.4.1
psutil==7.2.1
py==1.11.0
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
pyee==13.0.0
Pygments==2.19.2
PyMySQL==1.1.2
pypika-tortoise==0.6.3
pypiwin32==223
PySocks==1.7.1
pytest==8.4.2
pytest-base-url==2.1.0
pytest-playwright==0.6.2
python-dotenv==1.2.1
python-multipart==0.0.21
python-slugify==8.0.4
pytz==2025.2
pywin32==311
pywinauto==0.6.3
PyYAML==6.0.3
requests==2.32.5
retry==0.9.2
selenium==4.40.0
simple-tornado==0.2.2
simplejson==3.20.2
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
starlette==0.50.0
tabulate==0.9.0
text-unidecode==1.3
tidevice==0.12.11
tornado==6.5.4
tortoise-orm==0.25.3
tqdm==4.67.1
trio==0.32.0
trio-typing==0.10.0
trio-websocket==0.12.2
types-certifi==2021.10.8.3
types-urllib3==1.26.25.14
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
uvicorn==0.40.0
websocket-client==1.9.0
win32_setctime==1.2.0
wrapt==1.17.3
wsproto==1.3.2
yarl==1.22.0
zipp==3.23.0
